tool: Build auditor against Dafny 3.9.0
This works as-is, for the auditor, but results in many warnings
building the other tools. I believe that once we bring the
`compiler-bootstrap` branch up-to-date with a post-3.9.0 commit we'll
be able to depend on `DafnyRuntime` on NuGet and go back to
`-useRuntimeLib`. For the moment, though, code built with
`-useRuntimeLib` seems to depend on version 0.0.0.0 of `DafnyRuntime`.